DEV-1 refactoring fails
