+++ src/App.js
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
-import {Profile} from './Profile'
-import {Auth} from "./Auth";
-import {Maps} from "./Maps";
-import {Login} from "./Login";
-import './App.css';
+import {Profile} from './Profile/Profile'
+import {Auth} from "./Auth/Auth";
+import {Maps} from "./Maps/Maps";
+import {Login} from "./Login/Login";
+import {Header} from "./Header/Header";
 
 class App extends Component {
   state = {currentPage: 'login'}
@@ -13,42 +13,15 @@ class App extends Component {
   }
 
   PAGES = {
-    login: <Login singIn={() => {
-      this.navigateTo('maps')}} auth={() => {
-      this.navigateTo('auth')}}/>,
-    auth: <Auth auth={() => {
-      this.navigateTo('maps')}} login={() => {
-      this.navigateTo('login')}}/>,
+    login: <Login navigateTo={this.navigateTo} />,
+    auth: <Auth navigateTo={this.navigateTo}/>,
     maps: <Maps/>,
-    profile: <Profile/>
+    profile: <Profile navigateTo={this.navigateTo}/>
   }
 
   render() {
     return (<>
-      <header>
-        <nav>
-          <ul >
-            <li>
-              <button onClick={() => {
-                this.navigateTo('auth')
-              }}>Регистрация
-              </button>
-            </li>
-            <li>
-              <button onClick={() => {
-                this.navigateTo('maps')
-              }}>Maps
-              </button>
-            </li>
-            <li>
-              <button onClick={() => {
-                this.navigateTo('profile')
-              }}>Profile
-              </button>
-            </li>
-          </ul>
-        </nav>
-      </header>
+      <Header navigateTo={this.navigateTo}/>
       <main>
         <section>{this.PAGES[this.state.currentPage]}</section>
       </main>
