+++ src/Login/Login.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export const Login = ({singIn, auth}) => {
+export const Login = ({navigateTo}) => {
   return (
     <>
       <h1>Войти</h1>
@@ -11,11 +11,11 @@ export const Login = ({singIn, auth}) => {
         <input id='password' type='password' name='password' size='58' placeholder='Пароль*'/>
         <br/>
         <br/>
-        <button onClick={singIn}>Войти</button>
+        <button onClick={() => navigateTo('maps')}>Войти</button>
       </form>
       <br/>
       <br/>
-      <p>Новый пользователь? <button type='button' onClick={auth}>Зарегистрируйтесь</button></p>
+      <p>Новый пользователь? <button type='button' onClick={() => navigateTo('auth')}>Зарегистрируйтесь</button></p>
     </>
   )
 }
